Update edit backup

+++ backend/controllers/BackupController.php
@@ -2,14 +2,24 @@
 
 namespace backend\controllers;
 
+use backend\models\Restore;
+use backend\models\UploadBackup;
+use common\models\Image;
+use yii\web\Controller;
 use yii\base\Exception;
+use yii\data\ArrayDataProvider;
+use yii\web\HttpException;
+use yii\web\UploadedFile;
 
-class BackupController extends \yii\web\Controller
+class BackupController extends Controller
 {
-    public function actionIndex()
-    {
-        return $this->render('index');
-    }
+    public $tables = [];
+    public $fp ;
+    public $file_name;
+    public $_path = null;
+    public $back_temp_file = 'db_backup_';
+    public $view_table = [];
+    public $count = 0;
 
     protected function getPath()
     {
@@ -23,6 +33,7 @@ class BackupController extends \yii\web\Controller
         }
         return $this->_path;
     }
+
     public function execSqlFile($sqlFile)
     {
         $message = "ok";
@@ -49,21 +60,35 @@ class BackupController extends \yii\web\Controller
         $cmd = \Yii::$app->db->createCommand($sql);
         $table = $cmd->queryOne();
 
+        $is_view = false;
+
         $create_query = '';
 
         if (array_key_exists('Create Table', $table)) {
             $create_query = $table['Create Table'] . ';';
         } else if(array_key_exists('Create View', $table)) {
+            $is_view = true;
             $create_query .= $table['Create View'] . ';';
         }
 
-        $create_query  = preg_replace('/^CREATE TABLE/', 'CREATE TABLE IF NOT EXISTS', $create_query);
-        $create_query = preg_replace('/AUTO_INCREMENT\s*=\s*([0-9])+/', '', $create_query);
+        if (!$is_view) {
+            $create_query = preg_replace('/^CREATE TABLE/', 'CREATE TABLE IF NOT EXISTS', $create_query);
+            $create_query = preg_replace('/AUTO_INCREMENT\s*=\s*([0-9])+/', '', $create_query);
+        } else if ($is_view) {
+            $create_query = preg_replace('/^CREATE (.*) VIEW/', 'CREATE OR REPLACE VIEW', $create_query);
+        }
+
         if ( $this->fp)
         {
-            $this->writeComment('TABLE `'. addslashes ($tableName) .'`');
-            $final = 'DROP TABLE IF EXISTS `' .addslashes($tableName) . '`;'.PHP_EOL. $create_query .PHP_EOL.PHP_EOL;
-            fwrite ( $this->fp, $final );
+            if(!$is_view) {
+                $this->writeComment('TABLE `' . addslashes($tableName) . '`');
+                $final = 'DROP TABLE IF EXISTS `' . addslashes($tableName) . '`;' . PHP_EOL . $create_query . PHP_EOL . PHP_EOL;
+                fwrite($this->fp, $final);
+            } else if ($is_view) {
+                $this->view_table[$this->count]['view_name'] = 'VIEW `' . addslashes($tableName) . '`';
+                $this->view_table[$this->count]['schema'] = $create_query . PHP_EOL . PHP_EOL;
+                $this->count++;
+            }
         }
         else
         {
@@ -92,7 +117,7 @@ class BackupController extends \yii\web\Controller
             $values ="\n" . $valueString;
             if ($values != "")
             {
-                $data_string .= "INSERT INTO `$tableName` (`$items`) VALUES" . rtrim($values, ",") . ";;;" . PHP_EOL;
+                $data_string .= "INSERT INTO `$tableName` (`$items`) VALUES" . rtrim($values, ",") . ";" . PHP_EOL;
             }
         }
 
@@ -111,6 +136,7 @@ class BackupController extends \yii\web\Controller
             return $data_string;
         }
     }
+
     public function getTables($dbName = null)
     {
         $sql = 'SHOW TABLES';
@@ -118,6 +144,7 @@ class BackupController extends \yii\web\Controller
         $tables = $cmd->queryColumn();
         return $tables;
     }
+
     public function StartBackup($addcheck = true)
     {
         $this->file_name =  $this->path . $this->back_temp_file . date('Y.m.d_H.i.s') . '.sql';
@@ -135,15 +162,30 @@ class BackupController extends \yii\web\Controller
         }
         fwrite ( $this->fp, 'SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0;'.PHP_EOL );
         fwrite ( $this->fp, 'SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0;'.PHP_EOL );
-        fwrite ( $this->fp, '-- -------------------------------------------'.PHP_EOL );
+        fwrite ( $this->fp, 'SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE="TRADITIONAL,ALLOW_INVALID_DATES";'.PHP_EOL);
+        fwrite ( $this->fp, '-- -------------------------------------------'.PHP_EOL . PHP_EOL);
+
         $this->writeComment('START BACKUP');
+
+        $this->writeComment('Schema fsmdb');
+        fwrite ($this-> fp, 'DROP SCHEMA IF EXISTS `fsmdb` ;'.PHP_EOL.PHP_EOL);
+        $this->writeComment('Schema fsmdb');
+        fwrite($this->fp, 'CREATE SCHEMA IF NOT EXISTS `fsmdb` DEFAULT CHARACTER SET utf8 ;'.PHP_EOL);
+        fwrite($this->fp, 'USE `fsmdb` ;'.PHP_EOL.PHP_EOL);
+
         return true;
     }
+
     public function EndBackup($addcheck = true)
     {
+        foreach ($this->view_table as $view) {
+            $this->writeComment($view['view_name']);
+            fwrite ( $this->fp, $view['schema']);
+        }
         fwrite ( $this->fp, '-- -------------------------------------------'.PHP_EOL );
         fwrite ( $this->fp, 'SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS;'.PHP_EOL );
         fwrite ( $this->fp, 'SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS;'.PHP_EOL );
+        fwrite ( $this->fp, 'SET SQL_MODE=@OLD_SQL_MODE;'.PHP_EOL);
 
         if ( $addcheck )
         {
@@ -164,12 +206,16 @@ class BackupController extends \yii\web\Controller
 
     public function actionCreate()
     {
+        $flashError = '';
+        $flashMsg = '';
+
         $tables = $this->getTables();
 
         if(!$this->StartBackup())
         {
             //render error
-            \Yii::$app->user->setFlash('success', "Error");
+            $flashError = 'error';
+            $flashMsg = 'Some errors creating the file';
             return $this->render('index');
         }
 
@@ -183,7 +229,120 @@ class BackupController extends \yii\web\Controller
         }
         $this->EndBackup();
 
+        $flashError = 'success';
+        $flashMsg = 'The file was created !!!';
+        \Yii::$app->getSession()->setFlash($flashError, $flashMsg);
         $this->redirect(['index']);
+    }
+
+    public function actionDelete($file = null) {
+
+        $file = $_GET[0]['filename'];
+
+        if (isset($file)) {
+            $sqlFile = $this->getPath() . basename($file);
+            if (file_exists($sqlFile)) {
+                unlink($sqlFile);
+                $flashError = 'success';
+                $flashMsg = 'The file ' . $sqlFile . ' was successfully deleted.';
+            } else {
+                $flashError = 'error';
+                $flashMsg = 'The file ' . $sqlFile . ' was not found.';
+            }
+        } else {
+            $flashError = 'error';
+            $flashMsg = 'The file ' . $file . ' was not found.';
+        }
+        \Yii::$app->getSession()->setFlash($flashError, $flashMsg);
+        $this->redirect(['index']);
+    }
+
+    protected function getFileList()
+    {
+        $path = $this->getPath();
+        $list = [];
+        $list_files = glob($path .'*.sql');
+        if ($list_files )
+        {
+            $list = array_map('basename',$list_files);
+            sort($list);
+        }
+        return $list;
+    }
+
+    public function actionIndex()
+    {
+        $dataArray = [];
+
+        $list = $this->getFileList();
+        foreach ( $list as $id=>$filename )
+        {
+            $columns = [];
+            $columns['id'] = $id;
+            $columns['name'] = basename ( $filename);
+            $columns['size'] = filesize ( $this->getPath(). $filename);
+
+            $columns['create_time'] = date( 'Y-m-d H:i:s', filectime($this->getPath() .$filename) );
+            $columns['modified_time'] = date( 'Y-m-d H:i:s', filemtime($this->getPath() .$filename) );
+
+            $dataArray[] = $columns;
+        }
+
+        $dataProvider = new ArrayDataProvider(['allModels'=>$dataArray]);
+        return $this->render('index', array(
+            'dataProvider' => $dataProvider,
+        ));
+    }
+
+    public function actionDownload($filename = null)
+    {
+        $errors = [];
+        $filePath = $this->getPath() . basename($filename);
+
+        if (!is_file($filePath)) {
+            $errors[] = 'file not found';
+        }
+        if ($errors !== []) {
+            throw new HttpException(403, implode(', ', $errors));
+        }
+
+        \Yii::$app->response->sendFile($filePath, $filename);
+    }
+
+    public function actionRestore($file = null) {
+        $flashError = '';
+        $flashMsg = '';
+
+        $file = $_GET[0]['filename'];
+//
+//        $this->updateMenuItems();
+        $sqlFile = $this->path . basename($file);
+        if (isset($file)) {
+            $sqlFile = $this->path . basename($file);
+            $flashError = 'success';
+            $flashMsg = 'Looks like working :)';
+        } else {
+            $flashError = 'error';
+            $flashMsg = 'Problems with the file name';
+        }
+        $this->execSqlFile($sqlFile);
+
+        \Yii::$app->getSession()->setFlash($flashError, $flashMsg);
+        $this->redirect(array('index'));
+    }
+
+    public function actionUpload() {
+        $model = new UploadBackup();
+        if ($model->load(\Yii::$app->request->post())) {
+
+            $file = UploadedFile::getInstance($model, 'upload_file');
+            if ($file->saveAs($this->getPath() . $file)) {
+                // redirect to success page
+                return $this->redirect(['index']);
+            }
+        }
+
+        return $this->render('upload', ['model' => $model]);
     }
 
 }
